Updated all of the test usages of the HttpServletResponse.SC_* constants to compare them as Integers and not call the .toString() method on them.

+++ repose-aggregator/tests/functional-tests/src/integrationTest/groovy/features/filters/apivalidator/absolutePaths/ValdiatorAbsoluteWadlPathTest.groovy
@@ -85,7 +85,7 @@ class ValdiatorAbsoluteWadlPathTest extends ReposeValveTest {
 
         then: "request returns a 404 and and no error wadl error is thrown"
         wadlError.size() == 0
-        resp.getReceivedResponse().code == 404.toString()
+        resp.getReceivedResponse().code as Integer == 404.toString()
     }
 
     def "when loading validators on startup, it will fail when it cannot find a wadl when given an absolute path"() {
